CASSANDRA-15833: Bring back Cassandra 3.0 column filter behaviour when upgrading
CASSANDRA-10657 introduced changes in how the ColumnFilter is interpreted. This results in digest mismatch when querying incomplete set of columns from a table with consistency that requires reaching instances running pre CASSANDRA-10657 from nodes that include CASSANDRA-10657 (it was introduced in Cassandra 3.4).

The fix is to bring back the previous behaviour until there are no instances running pre CASSANDRA-10657 version.